Refactored the fetch to napster logic
